move auth to chm_utils.ext
